TankerLoadDetail - change integer to decimal and input type to number

+++ database/migrations/2020_12_26_002414_create_tanker_load_details_table.php
@@ -17,7 +17,7 @@ class CreateTankerLoadDetailsTable extends Migration
             $table->id();
             $table->bigInteger('tanker_load_id')->nullable();
             $table->bigInteger('product_id')->nullable();
-            $table->integer('quantity')->nullable();
+            $table->decimal('quantity', 8, 0)->nullable();
             $table->timestamps();
         });
     }
